feat(client): Implement Connection interceptors
